+++ src/ts/main.ts
@@ -163,11 +163,19 @@ export {
   SettingsPanelPageNavigatorButton,
   SettingsPanelPageNavigatorConfig,
 } from './components/settings/SettingsPanelPageNavigatorButton';
+export {
+  SettingsPanelPageNavigationItem,
+  SettingsPanelPageNavigationItemConfig,
+} from './components/settings/SettingsPanelPageNavigationItem';
 export { SettingsPanelSeparator, SettingsPanelSeparatorConfig } from './components/settings/SettingsPanelSeparator';
 export {
   PlayerInsightsPanel,
   PlayerInsightsPanelConfig,
 } from './components/panels/player-insights/PlayerInsightsPanel';
+export {
+  PlayerInfoSettingsPanelPage,
+  PlayerInfoSettingsPanelPageConfig,
+} from './components/settings/PlayerInfoSettingsPanelPage';
 export { InteractiveSettingsPanelItem } from './components/settings/InteractiveSettingsPanelItem';
 export { TouchControlOverlay, TouchControlOverlayConfig } from './components/overlays/TouchControlOverlay';
 export { CharacterEdgeColorSelectBox } from './components/settings/subtitlesettings/CharacterEdgeColorSelectBox';